+++ controller/staffUpdateController.js
@@ -433,100 +433,6 @@ const updateInsuranceDetails = catchAsync(async (req, res) => {
   }
 });
 
-const incometaxUpdate = catchAsync(async (req, res) => {
-  try {
-    const { phoneNumber, status, id } = req.body;
-
-    const incomeTaxAcknowledgement = req?.files?.incomeTaxAcknowledgement;
-    const computationFile = req?.files?.computationFile;
-
-    if(!req.files){
-      return res.status(400).json({ message: "Files not uploaded" });
-    }
-
-    if (!phoneNumber && !id) {
-      return res.status(400).json({ message: "input fields are required" });
-    }
-
-    const incomeTaxFilingDetail = defineIncomeTax();
-
-    const data = await incomeTaxFilingDetail.findOne({
-      where: {
-        phoneNumber: phoneNumber,
-        id: id,
-      },
-    });
-
-    if (!data) {
-      return res.status(404).json({
-        message: "Record not found",
-      });
-    }
-
-    const uploadFile = async (file) => {
-      if (file) {
-        try {
-          return await uploadBlob(file);
-        } catch (error) {
-          console.error(Error`uploading file ${file.name}:`, error);
-          throw new Error("File upload failed");
-        }
-      }
-      return null;
-    };
-
-    if (data.typeofTransaction === "salaried"){
-      console.log("Transaction", data.typeofTransaction)
-      totalAmount = 500;
-      franchiseCommission = 100;
-      HOCommission = 400;
-    }
-
-    if(data.typeofTransaction === "business"){
-      totalAmount = 5000;
-      franchiseCommission = 1000;
-      HOCommission = 4000;
-    }
-    else if(data.typeofTransaction === "capitalGain"){
-      totalAmount = 5000;
-      franchiseCommission = 1000;
-      HOCommission = 4000;
-    }
-
-    if(data.typeofTransaction === "other"){
-      totalAmount = 1500;
-      franchiseCommission = 300;
-      HOCommission = 1200;
-    }
-
-    const incomeTaxAcknowledgementUrl = await uploadFile(
-      incomeTaxAcknowledgement
-    );
-    const computationFileUrl = await uploadFile(computationFile);
-
-    const finalStatus = status === "completed" ? "completed" : "inProgress";
-
-    data.status = finalStatus;
-    data.franchiseCommission = franchiseCommission || data.franchiseCommission;
-    data.HOCommission = HOCommission || data.HOCommission;
-    data.totalAmount = totalAmount || data.totalAmount;
-    data.incomeTaxAcknowledgement = incomeTaxAcknowledgementUrl || data.incomeTaxAcknowledgement;
-    data.computationFile = computationFileUrl || data.computationFile;
-
-    await data.save();
-
-    return res.status(200).json({
-      message: "Success",
-      data,
-    });
-  } catch (error) {
-    console.log(error);
-    return res
-      .status(500)
-      .json({ message: "An error occurred error", error: error.message });
-  }
-});
-
 const updateBusBooking = catchAsync(async (req, res) => {
   try {
     let status = "inProgress"
@@ -620,4 +526,4 @@ const updateBusBooking = catchAsync(async (req, res) => {
   }
 });
 
-module.exports = { loanStatus, trainBookingUpdate, updatePanDetails,updateGstDetails,updateInsuranceDetails,incometaxUpdate,updateBusBooking };
+module.exports = { loanStatus, trainBookingUpdate, updatePanDetails,updateGstDetails,updateInsuranceDetails,updateBusBooking };
