+++ route/incomeTaxFilingRoute.js
@@ -1,11 +1,15 @@
 const express = require("express");
 const { verifyToken, verifyRefreshToken } = require("../utils/token");
-const createIncomeTaxFiling = require("../controller/incomeTaxFilingController");
+const {createIncomeTaxFiling, incometaxUpdate } = require("../controller/incomeTaxFilingController");
 
 const router = express.Router();
 
 router
   .route("/incomeTaxFiling")
   .post(verifyToken, verifyRefreshToken, createIncomeTaxFiling);
+  
+router
+  .route("/incometaxUpdate")
+  .put(verifyToken, verifyRefreshToken, incometaxUpdate)
 
 module.exports = router;
